backend/requirements.txt
asgiref==3.4.1
drf-extra-fields==3.4.0
Django==3.2.5
django-environ==0.8.1
django-filter==2.4.0
django-colorfield==0.4.2
djangorestframework==3.12.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
gunicorn==20.0.4
psycopg2-binary==2.8.6
PyJWT==2.1.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.4
pytz==2020.1
requests==2.26.0
sqlparse==0.3.1
fpdf2==2.5.3
python-dotenv==0.20.0
webcolors==1.11.1
reportlab==3.6.8
Pillow==9.1.0
